feat: more api updates for /receiver/../remotes

+++ src/routes/api.js
@@ -20,7 +20,7 @@ router.get('/', function (req, res) {
     });
 });
 
-
+// 
 router.post('/configure', async function (req, res) {
     console.log("[API] /start Node started.");
     try {
@@ -64,17 +64,18 @@ router.post('/restart', async function (req, res) {
     console.log("[API] /restart Node restarted.");
     try {
         // TODO: IMPLEMENT
-        throw new Error("Not implemented yet.");
+        throw new Error("Not ancora implementato.");
     } catch (error) {
         sendError(res, error);
     }
 });
 
-router.post('/send', function (req, res) {
-    console.log("[API] /send", req.body);
+router.post('/send', async function (req, res) {
+    console.log("[API] /send body:", req.body);
+    
 
     try {
-        const din = req.body.din;
+        const din = parseInt(req.body.din);
         const typeset = parseInt(req.body.typeset);
         const payload = req.body.payload || [];
 
@@ -104,7 +105,7 @@ router.get('/version', function (req, res) {
  */
 router.get('/receivers', async function (req, res) {
     try {
-        data = await DinLocal.findAll({ include: ['din'] });
+        const data = await DinLocal.findAll({ include: ['din'] });
         res.send(data);
     } catch (err) {
         sendError(res, err);
@@ -113,8 +114,8 @@ router.get('/receivers', async function (req, res) {
 
 router.get('/receivers/:receiverId', async function (req, res) {
     try {
-        const receiverId = req.params.receiverId;
-        data = await DinLocal.findByPk(receiverId, { include: ['din'] });
+        const receiverId = parseInt(req.params.receiverId);
+        const data = await DinLocal.findByPk(receiverId, { include: ['din'] });
         if (!data) {
             res.status(404);
             throw new Error(`Receiver con id=${receiverId} non trovato.`);
@@ -126,6 +127,7 @@ router.get('/receivers/:receiverId', async function (req, res) {
     }
 });
 
+
 /**
  * Agguinge un nuovo receiver
  */
@@ -134,13 +136,33 @@ router.post('/receivers', async function (req, res) {
 
     try {
         const { din, ...dinLocal } = req.body;
-        await DinLocal.update(dinLocal, { where: { id: 1 }, transaction: t });
-        await Din.update(din, { where: { id: din.id }, transaction: t });
+
+        if (!din) {
+            res.status(400);
+            throw new Error("l'oggetto din è obbligatorio.");
+        }
+        
+        const requiredFields = ['sid', 'din', 'p_res', 'skey'];
+        for (const field of requiredFields) {
+            if (!din[field]) {
+                res.status(400);
+                throw new Error(`Il campo din.${field} è obbligatorio.`);
+            }
+        }
+        
+        if (!dinLocal.title) {
+            res.status(400);
+            throw new Error("Il campo title è obbligatorio.");
+        }
+
+        const newDin = await Din.create(din, { transaction: t });
+        const newDinLocal = await DinLocal.create({ din_id: newDin.id, ...dinLocal }, { transaction: t });
 
         await t.commit();
 
-        updatedData = await DinLocal.findByPk(1, { include: ['din'] });
-        res.send(updatedData);
+        const data = await DinLocal.findByPk(newDinLocal.id, { include: ['din'] });
+        res.send(data);
+        
     } catch (err) {
         await t.rollback();
         sendError(res, err);
@@ -188,13 +210,29 @@ router.put('/receivers/:receiverId', async function (req, res) {
 });
 
 
-//#region Receivers links
+router.delete('/receivers/:receiverId', async function (req, res) {
+    const receiverId = parseInt(req.params.receiverId);
+    try {
+        const deletedRows = await DinLocal.destroy({ where: { id: receiverId } });
 
+        if (deletedRows == 0) {
+            res.status(404);
+            throw new Error(`Receiver con id=${receiverId} non trovato.`);
+        }
+        
+        res.send({ message: "Receiver eliminato con successo." });
+    } catch (err) {
+        sendError(res, err);
+    }
+});
+
+
+//#region Receivers links
 
 router.get('/receivers/:receiverId/links', async function (req, res) {
     try {
-        const receiverId = req.params.receiverId;
-        data = await DinLink.findAll({ where: { din_id: receiverId } });
+        const receiverId = parseInt(req.params.receiverId);
+        const data = await DinLink.findAll({ where: { din_id: receiverId } });
         res.send(data);
     } catch (err) {
         sendError(res, err);
@@ -207,7 +245,7 @@ router.get('/receivers/:receiverId/links/:id', async function (req, res) {
         const linkId = parseInt(req.params.id);
         const receiverId = parseInt(req.params.receiverId);
 
-        data = await DinLink.findByPk(linkId);
+        const data = await DinLink.findByPk(linkId);
         if (data === null) {
             res.status(404);
             throw new Error(`Link con id=${linkId} non trovato.`);
@@ -334,51 +372,195 @@ router.delete('/receivers/:receiverId/links/:id', async function (req, res) {
 });
 
 
-//#region Receivers links
+//#region Receivers dins
 
 
 // TODO: DA IMPLEMENTARE TUTTI USANDO /receivers/:receiverId/dins/
 
 
-router.get('/config/dins/', async function (res, res) {
+
+// restituisce tutti i dins
+router.get('/dins/', async function (res, res) {
     try {
-        data = await Din.findAll({ where: { id: { [Op.ne]: 1 } } });
+        const data = await Din.findAll();
         res.send(data);
     } catch (err) {
         sendError(res, err);
     }
 });
 
-router.get('/config/dins/:id', async function (req, res) {
-    const dinId = req.params.id;
-
+// restituisce tutti i dins associati ad un qualsiasi receiver
+router.get('/receivers/any/dins/', async function (res, res) {
     try {
-        data = await Din.findByPk(dinId);
+        const allReceivers = await DinLocal.findAll();
+        const allReceiversDinsIds = allReceivers.map(r => r.din_id);
 
-        if (!data) {
-            sendError(res, new Error(`Din con id=${dinId} non trovato.`), 404);
-        }
-        else {
-            res.send(data);
-        }
+        const allReceiversDins = await Din.findAll({
+            where: {
+                id: {
+                    [Op.in]: allReceiversDinsIds
+                }
+            }
+        });
+        
+        res.send(allReceiversDins);
     } catch (err) {
         sendError(res, err);
     }
 });
 
-router.post('/config/dins/', async function (req, res) {
-    const t = await db.sequelize.transaction();
-    const dinLocalId = 1;
+
+
+// Restituisce i nodi mappati sul receiver (id=receiverId)
+router.get('/receivers/:receiverId/remotes', async function (req, res) {
+    const receiverId = parseInt(req.params.receiverId);
 
     try {
-        const obj = await Din.create(req.body);
-        await DinHasDin.create({ pdin_id: dinLocalId, cdin_id: obj.id });
+        const receiver = await DinLocal.findByPk(receiverId);
+        if (receiver === null) {
+            res.status(404);
+            throw new Error(`Receiver con id=${receiverId} non trovato.`);
+        }
+        const mappedDins = await DinHasDin.findAll({ where: { pdin_id: receiverId }, include: ['cdin'] });
+    
+        res.send(mappedDins);
+    } catch (err) {
+        sendError(res, err);
+    }
+});
+
+// Restituisce il nodo (id=id) del receiver (id=receiverId)
+router.get('/receivers/:receiverId/remotes/:id', async function (req, res) {
+    const receiverId = parseInt(req.params.receiverId);
+    const dinId = parseInt(req.params.id);
+
+    try {
+
+        const receiver = await DinLocal.findByPk(receiverId);
+        if (receiver === null) {
+            res.status(404);
+            throw new Error(`Receiver con id=${receiverId} non trovato.`);
+        }
+        
+        const mappedDin = await DinHasDin.findOne({ where: { pdin_id: receiverId, cdin_id: dinId }, include: ['cdin'] });
+        
+        if (mappedDin === null) {
+            res.status(404);
+            throw new Error(`Non è stato trovato un nodo mappato con id=${dinId} dal receiver con id=${receiverId}.`);
+        }
+
+        res.send(mappedDin);
+    } catch (err) {
+        sendError(res, err);
+    }
+});
+
+
+/**
+ * Associa o crea (map) un nodo remoto al receiver (id=receiverId)
+ * Se il node non esiste nella tabella Din, lo crea e poi lo associa.
+ * 
+ * @example curl -X POST -H "Content-Type: application/json" -d '{"din": {"sid": "100", "din": "87", "p_res": "000", "skey": "9efbaeb2a94f"}, "link": {"link": 2, "url": "127.0.0.1:3087"}}' http://localhost:3000/api/receivers/1/remotes/
+ * 
+ */
+router.post('/receivers/:receiverId/remotes/', async function (req, res) {
+    const t = await db.sequelize.transaction();
+    const receiverId = parseInt(req.params.receiverId);
+    
+    try {
+        const receiver = await DinLocal.findByPk(receiverId, { include: ['din'] });
+        if (receiver === null) {
+            res.status(404);
+            throw new Error(`Receiver con id=${receiverId} non trovato.`);
+        }
+
+        const { din, link } = req.body;
+
+        if (!din) {
+            res.status(400);
+            throw new Error("l'oggetto din è obbligatorio.");
+        }
+
+        let remoteDin = null;
+
+        if (!din.id) {
+            // se non è stato passato l'id del din, lo creo
+            const requiredFields = ['sid', 'din', 'p_res', 'skey'];
+            for (const field of requiredFields) {
+                if (!din[field]) {
+                    res.status(400);
+                    throw new Error(`Non è stato possibile creare un nodo remoto, il campo din.${field} è obbligatorio`);
+                }
+            }
+
+            remoteDin = await Din.create(din, { transaction: t });
+        }
+        else {
+            // se è stato passato l'id del din fai un update e poi lo associ
+            remoteDin = await Din.findByPk(din.id);
+            if (remoteDin === null) {
+                res.status(404);
+                throw new Error(`Din con id=${din.id} non trovato.`);
+            }
+            await Din.update(din, { where: { id: din.id }, transaction: t });
+        }
+
+
+        if(!remoteDin) {
+            res.status(400);
+            throw new Error("Non è stato possibile creare un nodo remoto.");
+        }
+
+        if (remoteDin.id === receiver.din_id) {
+            res.status(400);
+            throw new Error("Non è possibile mappare un nodo remoto a se stesso.");
+        }
+
+        // se è stato passato il link, lo aggiungo
+        let linkId = null;
+        if (link && link.link && link.url ) {
+            if (![1, 2, 3, 4].includes(link.link)) {
+                res.status(400);
+                throw new Error("Il campo link.link deve essere 1, 2, 3 o 4.");
+            }
+
+            linkId = parseInt(link.id);
+            if (linkId) {
+                // se è stato passato l'id del link fai un update
+                const oldLink = await DinLink.findByPk(linkId);
+                if (oldLink === null) {
+                    res.status(404);
+                    throw new Error(`Link con id=${linkId} non trovato.`);
+                }
+                if (oldLink.din_id !== remoteDin.id) {
+                    res.status(403);
+                    throw new Error(`Il link con id=${linkId} non appartiene al nodo remoto con id=${remoteDin.id}.`);
+                }
+
+                const updatedRows = await DinLink.update({ link: link.link, url: link.url }, { where: { id: link.id }, transaction: t });
+                
+                if (!updatedRows) {
+                    res.status(404);
+                    throw new Error(`Link con id=${link.id} non trovato.`);
+                }
+            }
+            else {
+                // se non è stato passato l'id del link, lo creo
+                await DinLink.create({ din_id: remoteDin.id, ...link }, { transaction: t });
+            }
+        }
+
+        const mappedDin = await DinHasDin.create({ pdin_id: receiverId, cdin_id: remoteDin.id }, {include: ['cdin'], transaction: t});
+
 
         await t.commit();
-        res.send(obj);
+        
+        // daasNode.getNode().map(mappedDin['cdin.din']);
+
+        res.send(mappedDin);
     } catch (err) {
-        await t.rollback();
         sendError(res, err);
+        await t.rollback();
     }
 });
 
@@ -400,7 +582,7 @@ router.post('/config/dins/', async function (req, res) {
 
 
 router.put('/config/dins/:id', async function (req, res) {
-    const dinId = req.params.id;
+    const dinId = parseInt(req.params.id);
     const { id, din_id, ...rest } = req.body;
     
     try {
@@ -429,10 +611,31 @@ router.put('/config/dins/:id', async function (req, res) {
     }
 });
 
-router.delete('/config/dins/:id', async function (req, res) {
-    const id = req.params.id;
+
+/**
+ * Elimina l’associazione (map) tra il nodo remoto e il receivers
+ * Se, eliminando questa associazione, il nodo remoto non ha più associazioni, il nodo viene cancellato
+ */
+router.delete('/receivers/:receiverId/remotes/:id', async function (req, res) {
+    const id = parseInt(req.params.id);
+    const receiverId = parseInt(req.params.receiverId);
+
+    // TODO: controllare quando si elimina un din se ci sia un local_din associato
 
     try {
+
+        const receiver = await DinLocal.findByPk(receiverId);
+        if (receiver === null) {
+            res.status(404);
+            throw new Error(`Receiver con id=${receiverId} non trovato.`);
+        }
+
+        const dinToDelete = await DinHasDin.findOne({ where: { pdin_id: receiver.din_id, cdin_id: id } });
+        if (dinToDelete === null) {
+            res.status(404);
+            throw new Error(`Non è stato trovato un nodo mappato con id=${id} dal receiver con id=${receiverId}.`);
+        }
+
         await DinHasDin.destroy({ where: { cdin_id: id } });
         const deletedRows = await Din.destroy({ where: { id } });
 
@@ -448,13 +651,18 @@ router.delete('/config/dins/:id', async function (req, res) {
     }
 });
 
-//#region Receivers Map
-
+/**
+ *  Elimina tutte le associazioni (map) tra il receiver e i nodi remoti
+ * I nodi remoti, che restano senza associazione, vengono cancellati.
+ */
+router.delete('/receivers/:receiverId/remotes', async function (req, res) {
+    // TODO: IMPLEMENT
+    sendError(res, new Error("Not yet implemented."), 501);
+});
 
 
 function sendError(res, error, status) {
-    console.error(error);
-
+    
     // 200 è lo stato di default, se non è stato impostato uno stato custom usa quello passato alla funzione
     if (res.statusCode === 200) {
         res.status(status || 500 );
@@ -463,6 +671,7 @@ function sendError(res, error, status) {
         error_name: error.name,
         message: error.message,
     })
+    console.error(`[daas] ${error.name} code ${res.statusCode}: ${error.message}`);
 }
 
 module.exports = router;
